pyproject.toml
[project]
name = "add-keywords-with-gemini"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai",
    "osxphotos>=0.74.2",
    "ruff",
    "tqdm>=4.67.1",
]
